refactor(workflows): flatten workflows/__tests__ — rename two parallel suites (#2565 slice 5)
workflows/__tests__ had 2 files, both colliding with top-level
basenames but each covering distinctly broader scope than its
top-level counterpart:

- parallel-executor.test.ts (top: 227 lines, utilities like
  allSucceeded/getFailedSteps/withRetries/executeParallel) vs
  parallel-executor.test.ts (nested: 685 lines, class-based tests of
  TaskQueue + ExecutionPlanner + ParallelExecutor orchestration)

- workflow-parser.test.ts (top: 437 lines, parseYaml/parseJson/
  validateWorkflow) vs workflow-parser.test.ts (nested: 696 lines,
  parser + loadFile + DependencyGraph + Integration Tests)

Rename nested files with scope suffixes that reflect what they
actually cover; both top + nested suites preserved:

- nested parallel-executor → parallel-executor-orchestration.test.ts
- nested workflow-parser → workflow-parser-integration.test.ts

Fix ../X imports → ./X and ../../X dynamic imports → ../X. Remove
the empty __tests__/ directory.

All 1265 workflow tests still pass.
